Add Measurement Conversions

+++ app/src/main/java/com/kithcenette/kitchenette/Ingredients.kt
@@ -4,10 +4,10 @@ class Ingredients{
     var id: Int = 0
     var recipeID : Int = 0
     var foodID : Int = 0
-    var quantity: Int = 0
+    var quantity: Double = 0.0
     var measurement: String = ""
 
-    constructor(recipeID:Int, foodID:Int, quantity:Int, measurement:String){
+    constructor(recipeID:Int, foodID:Int, quantity:Double, measurement:String){
         this.recipeID=recipeID
         this.foodID=foodID
         this.quantity=quantity
